requirements.txt
decord==0.6.0
einops==0.3.0
imageio==2.9.0
numpy==1.24.2
opencv_python
pandas==2.0.0
Pillow==9.5.0
pytorch_lightning==1.9.3
PyYAML==6.0
setuptools==65.6.3
torch==2.1.0
torchvision
tqdm==4.65.0
transformers==4.45.2
moviepy
av
xformers==0.0.22.post7
gradio
timm
scikit-learn 
open_clip_torch==2.22.0
kornia
tensorboard
fairscale
omegaconf
deepspeed
hydra-core
peft==0.9
av==13.1.0

